requirements.txt
# Core FastAPI and web framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Database and ORM
sqlalchemy==2.0.42
asyncpg==0.30.0
alembic==1.16.4

# Data validation and settings
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1

# Authentication and security
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
email_validator==2.2.0
bcrypt==4.3.0
cryptography==45.0.6


# Async and utility libraries
anyio==4.10.0
async-timeout==5.0.1
greenlet==3.2.4
sniffio==1.3.1
exceptiongroup==1.3.0

# HTTP and networking
h11==0.16.0
httptools==0.6.4
idna==3.10

# Development and utilities
python-dotenv==1.1.1
click==8.2.1
watchfiles==1.1.0
websockets==15.0.1

# Logging and display
colorama==0.4.6
colorlog==6.9.0

# Template and markup
mako==1.3.10
markupsafe==3.0.2

# Configuration and typing
pyyaml==6.0.2
tomli==2.2.1
typing-inspection==0.4.1
typing-extensions==4.14.1

# Type annotations
annotated-types==0.7.0

# Cryptographic dependencies
cffi==1.17.1
pycparser==2.22
dnspython==2.7.0
ecdsa==0.19.1
pyasn1==0.6.1
rsa==4.9.1
six==1.17.0

# Email sending
sendgrid==6.12.4